// src/pages/modules/PMB/entrance/data/types.ts
export interface IEntrance {
  id_jalur_masuk: string
  id_satuan_organisasi: string
  nama_jalur: string
  slug: string
  url_pendaftaran: string
  is_status_tampil: boolean
  urutan: number
  created_at: string
  created_user: string
  updated_at: string
  updated_user: string
  nama_satuan_organisasi: string
  nama_user_created: string
  nama_user_updated: string
}